restructure cedalion.image_reco into cedalion.dot. sketch dot.ImageReco class
