+++ src/cedalion/dot/__init__.py
@@ -1,0 +1,3 @@
+from . forward_model import ForwardModel
+from . head_model import TwoSurfaceHeadModel
+from . image_reco import ImageReco, RegularizationParams, SpatialBasisParams
