+++ src/cedalion/dot/tissue_properties.py
